Renamed AssetJson to AssetJsonObject. Separeted out the Position class. Moved entities to Entities folder.
