+++ Simulator/AiR_Simulator/AiR_Simulator/AssetSimulator.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text.Json;
+using AiR_Simulator.Entities;
 
 namespace AssetDataSimulator
 {
