+++ Simulator/AiR_Simulator/AiR_Simulator/Entities/AssetJsonObject.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace AiR_Simulator.Entities
+{
+    public class AssetJsonObject
+    {
+        public int AssetId { get; set; }
+        public List<Position> Positions { get; set; }
+    }
+}
