+++ Simulator/AiR_Simulator/AiR_Simulator/Entities/Position.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AiR_Simulator.Entities
+{
+    public class Position
+    {
+        public double X { get; set; }
+        public double Y { get; set; }
+    }
+}
